ci: lint the whole workspace, and clear what that surfaced
`cargo clippy` ran without --workspace, so only the root crate was
checked. broker, risk and rebalancer had accumulated 141 findings.

46 of them were the 100_00 = $100.00 digit-grouping convention. The root
crate declares it crate-wide in src/lib.rs; the other two did not, and
repeating the attribute per file misses every integration test, since
each file under tests/ is its own crate root. Set once per crate in
[lints.clippy], which does reach them.

38 more were machine-applied with cargo clippy --fix -- mostly nested
if-let collapsing into let-chains, now that MSRV 1.88 allows them. That
includes the Binance idempotency path, so the five idempotency and
reconnect drills were run separately: all green. collapsible_if only
fires where there is no else branch, so the rewrites cannot change which
arm runs.

The rest were dead-code warnings on shared test mocks. mock_binance.rs
and mock_tws.rs are each pulled into four test binaries, so a method one
binary drives -- simulate_websocket_disconnect, simulate_partial_fill --
looks dead to the other three. Allowed at the module level rather than
deleted, since deleting would break the drills.

validate_mock_vs_paper.rs is a genuine gap rather than a lint: it fills
ValidationResult.test_case and .divergences and never reads them, so the
harness records data it does not assert on. Silenced with that noted, so
the scaffolding survives for a proper assertion pass.

+++ rebalancer/tests/verification_tests.rs
@@ -531,7 +531,7 @@ fn test_recovery_time_realistic_order_counts() {
 
         // Verify recovery completed
         assert_eq!(state.checkpoint, Checkpoint::RunCompleted);
-        assert_eq!(state.orders.len() as usize, order_count);
+        assert_eq!(state.orders.len(), order_count);
 
         // Verify recovery time is reasonable (< 5 seconds for 100 orders)
         // This is a generous threshold; actual time should be much faster
